updated LuaGlue to the latest version
Former-commit-id: 84991055456f56ae4e2eb1c66094a4cb57092e4f
Former-commit-id: 03d16699514e7c3618eebeb80f6f5718a8b38e07

+++ plugins/lua/lib/LuaGlue/include/LuaGlue/LuaGlueSymTab.h
@@ -169,7 +169,7 @@ class LuaGlueSymTab
 		
 		const Symbol &findSym(uint32_t idx)
 		{
-			if(idx < 0 || idx > (int)items.size())
+			if(idx < 0 || idx > items.size())
 			{
 				//printf("findSym(%i): not found\n", idx);
 				return nullSymbol;
